Remove bug that prevents different query
Previous this change after using the calculator and pressing '='
the app would ignore the first value previous entering a new operator.

F.e: if you did (2 + 2 =) and later you tried to do ( 3 - 7 =) the
output was -3 because app was not registering that 3 after the first
equal and considered the new query started with the previous result
and the minus 7.